core.py
import networkx as nx
import numpy as np
from sentence_transformers import SentenceTransformer
from faiss import IndexFlatL2

class UniversalStateMachine:
    def __init__(self):
        # Symbolic Knowledge Graph
        self.graph = nx.DiGraph()
        
        # Sub-Symbolic Embeddings
        self.encoder = SentenceTransformer('all-MiniLM-L6-v2')
        self.embedding_index = IndexFlatL2(384)  # Match encoder dimension
        
        # Node tracking
        self.node_ids = []
        self.node_embeddings = np.empty((0, 384))
        
    # Lattice Compilation Algorithm
    def add_data(self, text):
        # Extract entities and relationships (simplified)
        entities = self._extract_entities(text)
        embeddings = self.encoder.encode(entities)
        
        # Update graph and embeddings
        for entity, emb in zip(entities, embeddings):
            node_id = hash(entity)
            if node_id not in self.graph:
                self.graph.add_node(node_id, label=entity)
                self._update_embeddings(node_id, emb)
            
        # Add relationships (simplified heuristic)
        self._infer_relationships(entities)
    
    # Voyager Calibration Algorithm (Online Learning)
    def _update_embeddings(self, node_id, embedding):
        self.node_ids.append(node_id)
        self.node_embeddings = np.vstack([self.node_embeddings, embedding])
        self.embedding_index.add(np.array([embedding]))
    
    # Electron Synthesis Algorithm (Efficient Query)
    def query(self, question, depth=2):
        query_emb = self.encoder.encode([question])
        _, indices = self.embedding_index.search(query_emb, 3)
        
        # Graph traversal from nearest nodes
        results = set()
        for idx in indices[0]:
            if idx >= 0:
                start_node = self.node_ids[idx]
                for _, neighbor in nx.dfs_edges(self.graph, start_node, depth):
                    results.add(self.graph.nodes[neighbor]['label'])
        return list(results)[:5]
    
    # Helper Methods
    def _extract_entities(self, text):
        return list(set(text.split()))  # Simplified tokenization
    
    def _infer_relationships(self, entities):
        for i in range(len(entities)-1):
            a = hash(entities[i])
            b = hash(entities[i+1])
            self.graph.add_edge(a, b)

# Training and Testing
if __name__ == "__main__":
    # Initialize USM
    usm = UniversalStateMachine()

    # Train with sample data
    corpus = [
        "AI systems use algorithms for decision making.",
        "Transformers rely on attention mechanisms.",
        "Knowledge graphs represent structured information."
    ]

    for text in corpus:
        usm.add_data(text)

    # Test Query
    test_query = "What do AI systems use?"
    results = usm.query(test_query)
    print(f"Query Response: {results}")

    # Performance Metrics
    import time

    # Benchmark Inference Speed
    start_time = time.time()
    for _ in range(100):
        usm.query("machine learning models")
    latency = (time.time() - start_time)/100

    # Accuracy Test
    test_cases = [
        ("decision making", "What do AI systems use?"),
        ("structured information", "Knowledge graphs represent...")
    ]

    correct = 0
    for answer, question in test_cases:
        results = usm.query(question)
        correct += int(answer in results)

    print(f"\nPerformance Metrics:")
    print(f"Latency: {latency*1000:.2f}ms per query")
    print(f"Accuracy: {correct/len(test_cases)*100}%")